set putbnr_base

+++ ft_printf/libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: gde-souz <gde-souz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/19 13:14:02 by gde-souz          #+#    #+#             */
-/*   Updated: 2023/08/09 12:24:51 by gde-souz         ###   ########.fr       */
+/*   Updated: 2023/09/06 12:16:00 by gde-souz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (*s)
